Bug fixing: pom, chart printers

+++ src/main/java/ChartTypeXStringsYNumbers.java
@@ -53,7 +53,8 @@ public class ChartTypeXStringsYNumbers implements ChartType {
         for (String[] row : givenData) {
             for (String cell : row) {
                 if (isNumeric(cell)) {
-                    numbers.add(Integer.parseInt(cell));
+//                    numbers.add(Integer.parseInt(cell));
+                    numbers.add(Double.parseDouble(cell));
                 }
             }
         }
